The boot notification should be sent only when booting, not when connecting to the server
According to the standard, a boot notification should be sent only when booting, but here a boot notification was also sent when the connection to the server was disconnected and reconnected.
I solved this problem, which caused some test cases to fail, with this method.
Maybe someone else has a better way to do this.