[ChargePoint] Added protected section
This could be a quality of life as there is no need to directly inherit from IChargePoint only, and thus reimplement what has been done in ChargePoint when only 1 or 2 methods need to be override.

As ChargePoint implement most methods, it is easier to inherit from ChargePoint especially when there are changes between releases.

The protected section adds only getters that return reference, thus for pointers an exception is thrown if not yet initialized.

Signed-off-by: Thomas BRUNEL <thomas.brunel2@icloud.com>